default implementation for UpdateCallbacks and UpdateWorkflowNoRejection for ergonomics
